mappingLoader/requirements.txt
azure-functions
pandas
numpy
lxml
sqlalchemy
mysql.connector
load_dotenv